test push to test branch

+++ functions.h
@@ -51,7 +51,7 @@ public:
 
 void printProcesses(vector<Process*> processes) {
     cout << "Printing out processes vector" << endl;
-    for (int i = 0; i < processes.size(); i++) {
+    for (unsigned long i = 0; i < processes.size(); i++) {
         processes[i]->printProcess();
     }
 }
@@ -59,7 +59,7 @@ void printProcesses(vector<Process*> processes) {
 
 vector<Process *> processArrived(vector<Process*> processes, int t) {
     vector <Process*> arrivedProcesses;
-    for (int i = 0; i < processes.size(); i++) {
+    for (unsigned long i = 0; i < processes.size(); i++) {
         if (processes[i]->t_arrival == t) {
             arrivedProcesses.push_back(processes[i]);
         }
@@ -73,7 +73,7 @@ void printQueue (vector<Process*> processes) {
         cout << "empty]" << endl;
         return;
     }
-    for (int i = 0; i < processes.size(); i++) {
+    for (unsigned long i = 0; i < processes.size(); i++) {
         cout << processes[i]->name;
     }
     cout << "]" << endl;
@@ -104,10 +104,12 @@ int calc_tau(double prev_tau, int actual, double alpha)
 
 double calculateAverageCPUBurst(Process * p) {
     double total = 0.0;
-    for(int i = 0; i < p->CPUBursts.size(); i++) {
+    for(unsigned long i = 0; i < p->CPUBursts.size(); i++) {
         total += double(p->CPUBursts[i]);
     }
     //cout << total/p->num_bursts << endl;
     return total;
 }
 #endif //CPUSCHEDULING_FUNCTIONS_H
+
+
